+++ src/applications/vaos/components/AppointmentRequestListItem.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import AdditionalInfo from '@department-of-veterans-affairs/formation-react/AdditionalInfo';
 
-import { formatBestTimetoCall, sentenceCase } from '../utils/appointment';
+import { formatBestTimetoCallForList, sentenceCase } from '../utils/formatters';
 
 import { APPOINTMENT_STATUS, TIME_TEXT } from '../utils/constants';
 import AppointmentStatus from './AppointmentStatus';
@@ -137,7 +137,7 @@ export default class AppointmentRequestListItem extends React.Component {
                     {appointment.phoneNumber}
                     <br />
                     <span className="vads-u-font-style--italic">
-                      {formatBestTimetoCall(appointment.bestTimetoCall)}
+                      {formatBestTimetoCallForList(appointment.bestTimetoCall)}
                     </span>
                   </dd>
                 </dl>
